+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 - Added dirents_extended function and a find substitution proposal.
+
+- Created a Guided Distributor using a mapping file to map chunks to specific
+  nodes.
+
 ## [0.8.0] - 2020-09-15
 ## New
 - Both client library and daemon have been extended to support the ofi+verbs